TestParser/requirements.txt
Django==4.1
djangorestframework==3.13.1
requests~=2.28.1
beautifulsoup4~=4.11.1
python-dotenv==0.20.0
django-filter~=22.1
django-adminlte3==0.1.6
djoser==2.1.0